issue#42: Refactor as typescript src/users
